Bug 794940 - Visible white line in Nexus 7. r=mfinkle, a=akeybl

+++ mobile/android/base/BrowserToolbarBackground.java
@@ -61,7 +61,7 @@ public class BrowserToolbarBackground extends LinearLayout
 
         int width = getMeasuredWidth();
         int height = getMeasuredHeight();
-        float curve = height * 1.125f;
+        int curve = (int) (height * 1.125f);
 
         mPath.reset();
